files/requirements.txt
python-telegram-bot==20.4
aiosqlite
python-dotenv
